Remove the task that does update the `seen` status of notes. It sas redundant since threading is already handled in Volley. Use a simple static method for it.
